Making GREAT Progress
I've now gotten to the point where the UI of my landing page is 100% complete and the "Check Status" button works. I'm using the endpoint with Riot's API that allows me to check the server status. My goal is to now create a second page that will lookup the player and display stats from their last ten matches.